+++ airflow_indexima/operators/indexima.py
@@ -132,28 +132,33 @@ class IndeximaQueryRunnerOperator(IndeximaHookBasedOperator):
 class IndeximaLoadDataOperator(IndeximaHookBasedOperator):
     r"""Indexima load data operator.
 
-    Operations:
+    This operator perform this task:
 
         1. truncate target_table (false per default)
-        2. load source_select_query into target_table using redshift_user_name credential
-        4. commit/rollback target_table
+        2. load `source_select_query` into `target_table` using `redshift_user_name` credential
+        3. commit/rollback target_table
 
-    All fields ('target_table', 'load_path_uri', 'source_select_query', 'truncate_sql',
-    'format_query', 'prefix_query', 'skip_lines', 'no_check', 'limit', 'locale',
-    'pause_delay_in_seconds_between_query' ) support airflow macro.
+    Note:
+        All fields support airflow macro.
+        ('target_table', 'load_path_uri', 'source_select_query', 'truncate_sql',
+        'format_query', 'prefix_query', 'skip_lines', 'no_check', 'limit', 'locale',
+        'pause_delay_in_seconds_between_query' )
 
-    Syntax (see https://indexima.com/support/doc/v.1.7/Load_Data/Load_Data_Inpath.html)
-    ```sql
-    LOAD DATA INPATH 'path_of_the_data_source'
-    INTO TABLE my_data_space
-    [FORMAT 'separator' / ORC / PARQUET / JSON];
-    [PREFIX 'value1 \t value2 \t ... \t']
-    [QUERY "my_SQL_Query"]
-    [SKIP lines]
-    [NOCHECK]
-    [LIMIT n_lines]
-    [LOCALE 'FR']
-    ```
+
+    .. Syntax:
+        https://indexima.com/support/doc/v.1.7/Load_Data/Load_Data_Inpath.html
+
+    Example:
+        >>>
+        LOAD DATA INPATH 'path_of_the_data_source'
+        INTO TABLE my_data_space
+        [FORMAT 'separator' / ORC / PARQUET / JSON];
+        [PREFIX 'value1 \t value2 \t ... \t']
+        [QUERY "my_SQL_Query"]
+        [SKIP lines]
+        [NOCHECK]
+        [LIMIT n_lines]
+        [LOCALE 'FR']
 
     """
 
@@ -256,7 +261,7 @@ class IndeximaLoadDataOperator(IndeximaHookBasedOperator):
         """Generate 'load data' sql query.
 
         Returns:
-            (str): load data sql query
+            str: load data sql query
 
         """