+++ airflow_indexima/uri/factory.py
@@ -15,16 +15,14 @@ def define_load_path_factory(conn_id: str, decorator: ConnectionDecorator, facto
     """Create an uri factory function with UriFactory profile.
 
     Example:
-    ```
-    def my_decorator(conn:Connection) -> Connection:
-        ...
-        return conn
+        >>>def my_decorator(conn:Connection) -> Connection:
+            ...
+            return conn
 
-    decorated_redshift_uri_factory = define_load_path_factory(
-        decorator=my_decorator,
-        factory=get_redshift_load_path_uri
-        )
-    ```
+        decorated_redshift_uri_factory = define_load_path_factory(
+            decorator=my_decorator,
+            factory=get_redshift_load_path_uri
+            )
 
     Args:
         conn_id (str): connection identifier of data source
@@ -32,7 +30,7 @@ def define_load_path_factory(conn_id: str, decorator: ConnectionDecorator, facto
         factory (UriGeneratorFactory): uri decorated factory
 
     Returns:
-        (UriFactory): function used as a macro to get load uri path
+        UriFactory: function used as a macro to get load uri path
 
     """