+++ airflow_indexima/uri/jdbc.py
@@ -18,7 +18,7 @@ def get_jdbc_load_path_uri(jdbc_type: str, connection_id: str, decorator: Option
         decorator (Optional[ConnectionDecorator]): optinal connection decorator
 
     Returns:
-        (str) load path uri
+        str: load path uri
 
     """
     conn = BaseHook.get_connection(connection_id)
@@ -41,18 +41,16 @@ def get_jdbc_load_path_uri(jdbc_type: str, connection_id: str, decorator: Option
 def get_redshift_load_path_uri(connection_id: str, decorator: Optional[ConnectionDecorator] = None) -> str:
     """Return redshift load path uri from a connection_id.
 
-    Example:
-    ```
-        get_redshift_load_path_uri(connection_id='my_conn')
-        >> 'jdbc:redshift://my-db:5439/db_client?ssl=true&user=airflow-user&password=XXXXXXXX'
-    ```
+    Examples:
+        >>> get_redshift_load_path_uri(connection_id='my_conn')
+        'jdbc:redshift://my-db:5439/db_client?ssl=true&user=airflow-user&password=XXXXXXXX'
 
     Args:
         connection_id (str): source connection identifier
         decorator (Optional[ConnectionDecorator]): optinal connection decorator
 
     Returns:
-        (str) load path uri
+        str: load path uri
 
     """
     return get_jdbc_load_path_uri(jdbc_type='redshift', connection_id=connection_id, decorator=decorator)
@@ -61,18 +59,16 @@ def get_redshift_load_path_uri(connection_id: str, decorator: Optional[Connectio
 def get_postgresql_load_path_uri(connection_id: str, decorator: Optional[ConnectionDecorator] = None) -> str:
     """Return postgresql load path uri from a connection_id.
 
-    Example:
-    ```
-        get_postgresql_load_path_uri(connection_id='my_conn')
-        >> 'jdbc:postgresql://my-db:5432/db_client?ssl=true&user=airflow-user&password=XXXXXXXX'
-    ```
+    Examples:
+        >>> get_postgresql_load_path_uri(connection_id='my_conn')
+        'jdbc:postgresql://my-db:5432/db_client?ssl=true&user=airflow-user&password=XXXXXXXX'
 
     Args:
         connection_id (str): source connection identifier
         decorator (Optional[ConnectionDecorator]): optinal connection decorator
 
     Returns:
-        (str) load path uri
+        str: load path uri
 
     """
     return get_jdbc_load_path_uri(jdbc_type='postgresql', connection_id=connection_id, decorator=decorator)